+++ src/App.tsx
@@ -1,21 +1,18 @@
-import { Component } from 'react';
 import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
-import Header from './components/Header';
+import Main from './components/Main';
+import NotFoundPage from './components/NotFoundPage';
 
-class App extends Component {
-  render() {
-    return (
-      <BrowserRouter>
-        <>
-          <Routes>
-            <Route path="/" element={<Navigate to="/1" />}></Route>
-            <Route path="/:numPage" element={<Header />} />
-            <Route path="*" element={<h1>Not Found</h1>} />
-          </Routes>
-        </>
-      </BrowserRouter>
-    );
-  }
-}
-
+const App = () => {
+  return (
+    <BrowserRouter>
+      <>
+        <Routes>
+          <Route path="/" element={<Navigate to="/1" />}></Route>
+          <Route path="/:numPage" element={<Main />} />
+          <Route path="*" element={<NotFoundPage />} />
+        </Routes>
+      </>
+    </BrowserRouter>
+  );
+};
 export default App;
