--- src/components/Header.module.css
@@ -1,45 +1,0 @@
-.header {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: #62a77c;
-  height: 100px;
-}
-
-input[type='text'] {
-  margin: 10px;
-  display: block;
-  width: 50%;
-  height: calc(2.25rem + 2px);
-  padding: 0.375rem 0.75rem;
-  font-size: 1rem;
-  font-weight: 400;
-  line-height: 1.5;
-  color: #212529;
-  background-color: #fff;
-  background-clip: padding-box;
-  border: 1px solid #bdbdbd;
-  border-radius: 0.25rem;
-  outline-color: #62a77c;
-  transition:
-    border-color 0.15s ease-in-out,
-    box-shadow 0.15s ease-in-out;
-}
-
-button {
-  background-color: #fff;
-  height: calc(2.25rem + 2px);
-  padding: 0.375rem 0.75rem;
-  font-size: 1.2rem;
-  line-height: 1.2;
-  color: #212529;
-  font-weight: 600;
-  background-clip: padding-box;
-  border: 1px solid #bdbdbd;
-  border-radius: 0.25rem;
-  cursor: pointer;
-}
-
-button:hover {
-  opacity: 0.9;
-}
