+++ src/components/Main.module.css
@@ -1,34 +1,47 @@
-.main {
+.header {
   display: flex;
-  flex-flow: wrap;
-  justify-content: space-around;
+  align-items: center;
+  justify-content: center;
+  background: #62a77c;
+  height: 100px;
+  margin-bottom: 20px;
 }
 
-.pokemon {
-  width: 220px;
-  padding: 10px;
-  border-radius: 5px;
+input[type='text'] {
   margin: 10px;
-  text-align: center;
-  box-shadow:
-    2px 5px 15px 0px #0b325e,
-    5px 5px 15px 5px rgba(0, 0, 0, 0);
+  display: block;
+  width: 50%;
+  height: calc(2.25rem + 2px);
+  padding: 0.375rem 0.75rem;
+  font-size: 1rem;
+  font-weight: 400;
+  line-height: 1.5;
+  color: #212529;
+  background-color: #fff;
+  background-clip: padding-box;
+  border: 1px solid #bdbdbd;
+  border-radius: 0.25rem;
+  outline-color: #62a77c;
+  transition:
+    border-color 0.15s ease-in-out,
+    box-shadow 0.15s ease-in-out;
 }
 
-.pokemon img {
-  width: 200px;
-  height: 200px;
+button {
+  background-color: #fff;
+  height: calc(2.25rem + 2px);
+  padding: 0.375rem 0.75rem;
+  font-size: 1.2rem;
+  line-height: 1.2;
+  color: #212529;
+  font-weight: 600;
+  background-clip: padding-box;
+  border: 1px solid #bdbdbd;
+  border-radius: 0.25rem;
+  cursor: pointer;
+  width: 120px;
 }
 
-.pokemon p {
-  margin: 5px;
-  text-align: center;
-  font-size: 18px;
-}
-
-.errName {
-  width: 100%;
-  margin-top: 20px;
-  font-size: 36px;
-  text-align: center;
+button:hover {
+  opacity: 0.9;
 }
