+++ src/components/NotFoundPage.module.css
@@ -1,0 +1,31 @@
+.not_found {
+  font-size: 30px;
+  font-weight: 700;
+  text-align: center;
+  margin-top: 20px;
+}
+
+.not_found span {
+  display: block;
+  margin-bottom: 10px;
+}
+
+.not_found a {
+  background-color: #fff;
+  height: calc(2.25rem + 2px);
+  padding: 0.375rem 0.75rem;
+  font-size: 1.2rem;
+  line-height: 1.2;
+  color: #212529;
+  font-weight: 600;
+  background-clip: padding-box;
+  border: 1px solid #bdbdbd;
+  border-radius: 0.25rem;
+  cursor: pointer;
+  width: 120px;
+  text-decoration: none;
+}
+
+.not_found a:hover {
+  opacity: 0.9;
+}
