+++ src/components/Pagination.module.css
@@ -1,26 +1,34 @@
-.pagination {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
 .pagination a {
   width: 30px;
   height: 30px;
   display: block;
-  margin: 10px;
+  margin: 7px;
   border-radius: 3px;
   background: #f7f7f7;
   text-decoration: none;
   padding: 5px 0px;
   text-align: center;
   cursor: pointer;
+  font-weight: 600;
 }
 
 .pagination a:hover {
-  color: red;
+  opacity: 0.7;
 }
 
 .pagination span {
   cursor: pointer;
+}
+
+.pagination_item {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-bottom: 20px;
+}
+
+.current_element {
+  margin-top: 10px;
+  text-align: center;
+  font-weight: 600;
 }
