+++ src/components/Pagination.tsx
@@ -10,6 +10,9 @@ interface Current {
 const Pagination = (props: Current) => {
   const currnPokemoOnPage = 20;
   const maxItemPaginatin = props.currentAllPokemons / currnPokemoOnPage;
+  const numPaginationPageFromLocalStorage = Number(
+    localStorage.getItem('numPaginationPage')
+  );
 
   const currentPages: Array<number> = [];
   for (let i = 0; i < maxItemPaginatin; i++) {
@@ -19,34 +22,45 @@ const Pagination = (props: Current) => {
   function countingItemPagination(numItemPagination: {
     currentTarget: { innerHTML: unknown };
   }) {
+    const numItemPagePagination = Number(
+      numItemPagination.currentTarget.innerHTML
+    );
+    localStorage.setItem('numPaginationPage', `${numItemPagePagination}`);
     props.clickOnItemPagination(
-      Number(numItemPagination.currentTarget.innerHTML) * currnPokemoOnPage -
-        currnPokemoOnPage
+      numItemPagePagination * currnPokemoOnPage - currnPokemoOnPage
     );
   }
 
   function creatItemPagination() {
     return (
       <>
-        {currentPages.map((numPage) => {
-          return (
-            <NavLink
-              to={`/${numPage}`}
-              key={numPage}
-              onClick={countingItemPagination}
-              end
-            >
-              {numPage}
-            </NavLink>
-          );
-        })}
+        {!(currentPages.length === 1) &&
+          currentPages.map((numPage) => {
+            return (
+              <NavLink
+                to={`/${numPage}`}
+                key={numPage}
+                onClick={countingItemPagination}
+                end
+              >
+                {numPage}
+              </NavLink>
+            );
+          })}
+
+        {}
       </>
     );
   }
 
   return (
     <>
-      <div className={styles.pagination}>{creatItemPagination()}</div>
+      <div className={styles.pagination}>
+        <div className={styles.current_element}>{`${
+          numPaginationPageFromLocalStorage * 20 - 19
+        } to ${numPaginationPageFromLocalStorage * 20}`}</div>
+        <div className={styles.pagination_item}>{creatItemPagination()}</div>
+      </div>
     </>
   );
 };
