+++ src/components/Pokemons.module.css
@@ -1,0 +1,34 @@
+.pokemons {
+  display: flex;
+  flex-flow: wrap;
+  justify-content: center;
+}
+
+.pokemon {
+  width: 220px;
+  padding: 10px;
+  border-radius: 5px;
+  margin: 10px;
+  text-align: center;
+  box-shadow:
+    2px 5px 15px 0px #0b325e,
+    5px 5px 15px 5px rgba(0, 0, 0, 0);
+}
+
+.pokemon img {
+  width: 200px;
+  height: 200px;
+}
+
+.pokemon p {
+  margin: 5px;
+  text-align: center;
+  font-size: 18px;
+}
+
+.errName {
+  width: 100%;
+  margin-top: 20px;
+  font-size: 36px;
+  text-align: center;
+}
